Remove IPCBufferToken and IPCBufferGuard
In order to simplify our public APIs, we opt to put
more trust in our IPC implementation's care around
the shared buffer resource, and supply internal convenience
helpers to reduce the odds of mechanical errors.

The primary cost of this is performance -- copying to and
from the IPC Buffer is now the norm even for the sync
Call + ReplyRecv fast path.