Make CNodeSlots be regular Caps, fix generate_self_reference
Previously, CNodeSlots were indexed only be the Role of the CNode whose slots
were being represented. This didn't allow transferring slots to a child process;
to enable this, change it to CNodeSlotsData and make it a CapType.

This also updates all CNodeSlots call sites to use the encapsulated
destructor (elim) and private constructor, removing boilerplate and
outside-the-abstraction code, reducing the possibility of mistaken resource
leaks.