update the stack and args setup w/r/t aarch64 docs
To the best of my ability translating
http://infocenter.arm.com/help/topic/com.arm.doc.ihi0055b/IHI0055B_aapcs64.pdf
5.4.2 stage C: My understanding here is that if the params do not fit
in the args, then the whole thing ought to go into the stack, aligned
at (guessing here) 8 bytes.

The document states to use the "natural alignment" of the "composite
type", this means to use the natural alignment of the largest
"fundemental type" member of the composite type.