Replace smart_alloc uuid use with a counter
The uuid dependency in smart_alloc is problematic with cargo's dependency
resolution system. If you need uuid at runtime in some other crate in the same
dependency tree, then they both get the 'std' feature.

This replaces the use of uuid with a global incrementing counter. This is
exactly the strategy used in Clojure's 'gensym'. It also renames make_ident to
gensym, since that's the common name for this operation.