Allow multipage queues, move capacity to data
This changes IPC queues (in multi_consumer) in two ways:

1. Queues may now span multiple pages. To support this, the Consumer
constructors and add_queue methods no take an additional QueueSizeBits
type paramter. No relationship between the length of the queue, the size
of the queued item and the size of the buffer is presently checked at
compile time (as we lack TypenumSized). A runtime assertion is provided
during setup.

2. Queue capacity is no longer reflected in the type of the producer,
   the conumser, or the underlying ArrayQueue.

Fixes: #57, #58